+++ src/elements/GeoEditableInput/GeoEditableInput.vue
@@ -11,6 +11,7 @@
       ref="dropdown"
       :opened="showButtons"
       :popup-class="['geo-editable-input__popup', `geo-editable-input__popup--${inputType}`, popupClass]"
+      :reposition-on-update="true"
       preferred-x-axis-position="right"
       @click-outside="handleClickOutside($event)"
     >
@@ -54,7 +55,7 @@
 </template>
 
 <script>
-import enumPropertyFactory from '../../utils/enumPropertyFactory'
+import { enumPropertyFactory } from '../../utils/enumPropertyFactory'
 import { VARIANTS, TYPES } from './GeoEditableInput.constants'
 
 /**
@@ -142,16 +143,12 @@ export default {
      * - `VARIANTS.table`
      * - `VARIANTS.normal`
      */
-    inputType: {
-      type: String,
-      default: function () {
-        return VARIANTS.table
-      },
-      validator: function (value) {
-        // The value must match one of these strings
-        return value in VARIANTS
-      }
-    },
+    inputType: enumPropertyFactory({
+      componentName: 'GeoEditableInput',
+      propertyName: 'inputType',
+      enumDictionary: VARIANTS,
+      defaultValue: VARIANTS.table
+    }),
 
     /**
      * Class or classes that will be added to the popup element