+++ src/elements/GeoInput/GeoInputMessage.vue
@@ -16,6 +16,7 @@
 
 <script>
 import { VARIANTS } from './GeoInputMessage.constants'
+import { enumPropertyFactory } from '../../utils/enumPropertyFactory'
 
 /**
  * `GeoInputMessage` is component designed to fit nicely as a help message for
@@ -40,17 +41,12 @@ export default {
      * See [Component Constants](/docs/guides/using-constants) for more info on how
      * to use those constants in your code.
      */
-    variant: {
-      type: String,
-      validator (value) {
-        if (!value) return true
-        if (value in VARIANTS) return true
-
-        const supportedValues = Object.values(VARIANTS).map(i => `«${i}»`).join(', ')
-        console.warn(`GeoInputMessage [component] :: Unsupported value («${value}») for «variant» property. Use one of ${supportedValues}`)
-        return false
-      }
-    }
+    variant: enumPropertyFactory({
+      componentName: 'GeoInputMessage',
+      propertyName: 'variant',
+      enumDictionary: VARIANTS,
+      checkUndefined: true
+    })
   }
 }
 </script>