+++ src/elements/GeoPill/GeoPill.vue
@@ -5,6 +5,8 @@
 </template>
 
 <script>
+import { enumPropertyFactory } from '../../utils/enumPropertyFactory'
+
 const VARIANTS = {
   default: 'default',
   light: 'light',
@@ -38,13 +40,12 @@ export default {
      * - `grey`
      * - `opaque`
      */
-    variant: {
-      type: String,
-      default: VARIANTS.default,
-      validator (value) {
-        return value in VARIANTS
-      }
-    }
+    variant: enumPropertyFactory({
+      componentName: 'GeoPill',
+      propertyName: 'variant',
+      enumDictionary: VARIANTS,
+      defaultValue: VARIANTS.default
+    })
   }
 }
 </script>