+++ test/unit/specs/elements/GeoTrimmedContent.spec.js
@@ -6,8 +6,6 @@ import GeoTooltip from '@/elements/GeoTooltip/GeoTooltip.vue'
 const localVue = createLocalVue()
 localVue.component('geo-trimmed-content', GeoTrimmedContent)
 
-jest.mock('vue-directive-tooltip')
-
 describe('GeoTrimmedContent', () => {
   it('Should render content', function () {
     const wrapper = mount(GeoTrimmedContent, {
@@ -22,5 +20,47 @@ describe('GeoTrimmedContent', () => {
     expect(trimmedContent.exists()).toBe(true)
     expect(wrapper.find('.my-content').exists()).toBe(true)
     expect(wrapper.find('.my-content').text()).toEqual('Custom content')
+  })
+
+  it('Should complain if tooltipPosition is invalid', function () {
+    const consoleWarnSpy = jest.spyOn(global.console, 'warn').mockImplementation(() => { })
+    const consoleErrorSpy = jest.spyOn(global.console, 'error').mockImplementation(() => { })
+
+    const wrapper = mount(GeoTrimmedContent, {
+      propsData: {
+        tooltipPosition: 'invalid position'
+      },
+      slots: {
+        default: '<div class="my-content">Custom content</div>'
+      },
+      stubs: {
+        GeoTooltip
+      }
+    })
+    const trimmedContent = wrapper.find('.geo-trimmed-content')
+    expect(trimmedContent.exists()).toBe(true)
+    expect(consoleErrorSpy).toHaveBeenCalled()
+    expect(consoleWarnSpy).toHaveBeenCalled()
+  })
+
+  it('Should complain if tooltipAlignment is invalid', function () {
+    const consoleWarnSpy = jest.spyOn(global.console, 'warn').mockImplementation(() => { })
+    const consoleErrorSpy = jest.spyOn(global.console, 'error').mockImplementation(() => { })
+
+    const wrapper = mount(GeoTrimmedContent, {
+      propsData: {
+        tooltipAlignment: 'invalid alignment'
+      },
+      slots: {
+        default: '<div class="my-content">Custom content</div>'
+      },
+      stubs: {
+        GeoTooltip
+      }
+    })
+    const trimmedContent = wrapper.find('.geo-trimmed-content')
+    expect(trimmedContent.exists()).toBe(true)
+    expect(consoleErrorSpy).toHaveBeenCalled()
+    expect(consoleWarnSpy).toHaveBeenCalled()
   })
 })